% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{pAlg-class}
\alias{pAlg}
\alias{pAlg-class}
\title{Class that represents a polynomial function}
\description{
Class that represents a polynomial function
}
\section{Slots}{

\describe{
\item{\code{tab}}{Array keeping all the information about the function (see below)}
}}

