% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polyAlgebra.R
\docType{package}
\name{polyAlgebra-package}
\alias{polyAlgebra-package}
\title{Package for analitical calculations on multiple variable polynomials}
\description{
Package for analitical calculations on multiple variable polynomials
}

